# highlight-erp/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Highlight ERP API Documentation",
        "description": "Документация API для проекта Highlight ERP",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Основной API сервер"
        }
    ],
    "paths": {
...
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Администратор - Сотрудники"
                ],
                "summary": "Получение списка всех сотрудников",
                "operationId": "9ded17315e13756cd593b5840cd39a26",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Администратор - Сотрудники"
                ],
                "summary": "Создание нового сотрудника",
                "operationId": "f6f90d5fbbe24f66eb2b43901e0d3f38",
                "requestBody": {
                    "description": "Данные нового сотрудника",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Сотрудник успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/{user}": {
            "get": {
                "tags": [
                    "Администратор - Сотрудники"
                ],
                "summary": "Получение информации о конкретном сотруднике",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID сотрудника",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Сотрудник не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Администратор - Сотрудники"
                ],
                "summary": "Обновление данных сотрудника",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID сотрудника",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Обновленные данные сотрудника",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Сотрудник успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Сотрудник не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации или бизнес-логики"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Администратор - Сотрудники"
                ],
                "summary": "Удаление сотрудника",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID сотрудника",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Сотрудник успешно удален"
                    },
                    "404": {
                        "description": "Сотрудник не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Аутентификация пользователя",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "password"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "format": "phone",
                                        "example": "79990001122"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная аутентификация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|aBcDeFgHiJkLmNoPqRsTuVwXyZ"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Выход из системы",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Успешный выход",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Вы успешно вышли из системы."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не аутентифицирован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreUserRequest": {
                "title": "StoreUserRequest",
                "description": "Данные для создания нового сотрудника",
                "required": [
                    "last_name",
                    "first_name",
                    "position",
                    "phone",
                    "password"
                ],
                "properties": {
                    "last_name": {
                        "description": "Фамилия",
                        "type": "string",
                        "example": "Петров"
                    },
                    "first_name": {
                        "description": "Имя",
                        "type": "string",
                        "example": "Петр"
                    },
                    "patronymic": {
                        "description": "Отчество",
                        "type": "string",
                        "example": "Петрович",
                        "nullable": true
                    },
                    "position_id": {
                        "description": "ID должности",
                        "type": "integer",
                        "example": "1"
                    },
                    "phone": {
                        "description": "Номер телефона (логин)",
                        "type": "string",
                        "example": "79991234502"
                    },
                    "password": {
                        "description": "Пароль (минимум 8 символов)",
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    },
                    "documents": {
                        "description": "Массив ID документов для назначения",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            },
            "UpdateUserRequest": {
                "title": "UpdateUserRequest",
                "description": "Данные для обновления сотрудника",
                "required": [
                    "last_name",
                    "first_name",
                    "position",
                    "phone"
                ],
                "properties": {
                    "last_name": {
                        "description": "Фамилия",
                        "type": "string",
                        "example": "Петров"
                    },
                    "first_name": {
                        "description": "Имя",
                        "type": "string",
                        "example": "Петр"
                    },
                    "patronymic": {
                        "description": "Отчество",
                        "type": "string",
                        "example": "Петрович",
                        "nullable": true
                    },
                    "position": {
                        "description": "Должность",
                        "type": "string",
                        "example": "Старший повар"
                    },
                    "phone": {
                        "description": "Номер телефона (логин)",
                        "type": "string",
                        "example": "79991234502"
                    },
                    "password": {
                        "description": "Пароль (минимум 8 символов)",
                        "type": "string",
                        "format": "password",
                        "example": "password123",
                        "nullable": true
                    },
                    "documents": {
                        "description": "Массив ID документов для назначения",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "Модель пользователя",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "last_name": {
                        "description": "Фамилия",
                        "type": "string",
                        "example": "Иванов"
                    },
                    "first_name": {
                        "description": "Имя",
                        "type": "string",
                        "example": "Иван"
                    },
                    "patronymic": {
                        "description": "Отчество",
                        "type": "string",
                        "example": "Иванович",
                        "nullable": true
                    },
                    "position": {
                        "description": "Должность",
                        "type": "string",
                        "example": "Главный администратор"
                    },
                    "phone": {
                        "description": "Номер телефона (логин)",
                        "type": "string",
                        "example": "79990001122"
                    },
                    "role": {
                        "description": "Роль пользователя",
                        "type": "string",
                        "enum": [
                            "admin",
                            "employee"
                        ],
                        "example": "admin"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введите Bearer токен для доступа к защищенным эндпоинтам",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}